[BUGFIX] Use callable instead of closure in the event runner (#6)
Because a closure can't be serialized, an exception would be thrown in
some cases when TYPO3 slot dispatcher is somehow serialized (e.g. when
put in cache for some reason).

With this patch, the API does not work with closures anymore, and uses a
basic callable instead (resulting in the same process in the end).