[FEATURE] Resolve mail template path from event identifier
The template file for a mail notification is not forced anymore to a
static path.

It now has two dynamic resolving methods:

- The identifier of both the dispatched event and its group are
  sanitized and used to guess a potential path.

  For instance, the template path for the event `myEvent` from the group
  `my_company` will be located at `MyCompany/MyEvent.html`.

- If the first method has no result, the template `Default` is used (one
  is provided by NotiZ but can be overridden).

To handle this new feature, a new definition entry has been added:
`notifications.entityEmail.settings.view.templateRootPaths` which can be
filled the classical way when working with Fluid templates.

Finally, the definition changed for both `layoutRootPaths` and
`partialRootPaths`: the folder `Mail` has been appended to the paths, to
mark it as the entry point for the template files.