[FEATURE] Introduce slots for dynamic body in mail notification (#35)
With this feature, the body section of the mail notification can now be
composed of dynamic fields, that are managed by so-called "slots". This
allows editors to handle several sections of the mail body, while the
templating itself stays in the Fluid view.

The slots can be registered in the template of the mail, in a Fluid
section named `Slots`. Two view-helpers are provided out of the box:

- `<nz:slot.text>` will register a new textarea field.
- `<nz:slot.input>` will register a new text-input field.

Because the registration happens in Fluid, basic operations like loops
and conditions can be used.

Slots can then be rendered within the template by using the following
view-helper: `<nz:slot.render>`. Additional markers may be added to the
slot by using the arguments `markers`.

See documentation for more information about this feature.

---

Closes #26