[FEATURE] Enhance `slot.render` view-helper
This view-helper can now use two new features, for a total of three ways
to render a slot.

Inline
------

The processed slot value will be returned.

```html
<nz:slot.render name="MySlot"
                markers="{foo: 'bar'}" />
```

Conditional
-----------

Can be used to check whether the slot exists, and do something if it
doesn't. When using this way, a variable `slotValue` becomes accessible
within the view-helper, that contains the processed value of the slot.
However, this variable is filled only when the slot exists and can be
processed.

```html
<nz:slot.render name="SomeOptionalSlot">
    <f:then>
        {slotValue -> f:format.html()}
    </f:then>
    <f:else>
        Some default value
    </f:else>
</nz:slot.render>
```

Wrapping
--------

You may need to add HTML around the slot value only when the slot
exists.

```html
<nz:slot.render name="SomeOptionalSlot">
    <hr />
    <div class="some-class">
        {slotValue}
    </div>
</nz:slot.render>
```