[FEATURE] Introduce notification manager backend module (#135)
This module gives access to different views, where notifications and
events can be managed easily. The usefulness is to centralise every
notification-related operation in a single place.

Editors can now create and edit notifications in a very simple and
intuitive way. They can also see a detailed view of each existing
record, including a preview area.

It is advised to configure editors right access to the new module, as
well as their abilities to create/see every type of notification.